[metasrv] refactor: move raft-log impl into metasrv/raft/log
