+++ metasrv/src/raft/log/raft_log_test.rs
@@ -18,9 +18,9 @@ use async_raft::raft::EntryPayload;
 use async_raft::LogId;
 use common_runtime::tokio;
 
-use crate::meta_service::raft_log::RaftLog;
 use crate::meta_service::Cmd;
 use crate::meta_service::LogEntry;
+use crate::raft::log::RaftLog;
 use crate::tests::service::new_sled_test_context;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
