+++ metasrv/src/raft/state_machine/mod.rs
@@ -23,6 +23,8 @@ mod placement_test;
 #[cfg(test)]
 mod state_machine_test;
 
+pub use applied_state::AppliedState;
+pub use placement::Placement;
 pub use sm::Node;
 pub use sm::Replication;
 pub use sm::SerializableSnapshot;
